relocate cachekey test assertions
